backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
python-dotenv
supabase
google-generativeai
Werkzeug
